requirements.txt
absl-py==2.1.0
addict==2.4.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
grpcio==1.68.1
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.1
tb-nightly==2.19.0a20241220
tensorboard-data-server==0.7.2
tifffile==2024.12.12
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.52
ultralytics-thop==2.0.13
urllib3==2.2.3
Werkzeug==3.1.3
yapf==0.43.0
